+++ tests/ParseExpressionTest.php
@@ -390,6 +390,15 @@ QRY
 
     public function testGenericOperator(): void
     {
+        $deprecations = '';
+        set_error_handler(
+            function (int $errno, string $errstr) use (&$deprecations) {
+                $deprecations .= $errstr;
+                return true;
+            },
+            \E_USER_DEPRECATED
+        );
+
         $list = $this->parser->parseExpressionList(<<<QRY
     w # @ v ? u, q !, ! q, r operator(blah.###) s
 QRY
@@ -427,6 +436,9 @@ QRY
             ]),
             $list
         );
+        restore_error_handler();
+
+        $this::assertStringContainsString('Postfix operators', $deprecations);
     }
 
     /**